requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
Django==6.0
docstring_parser==0.17.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-aiplatform==1.132.0
google-cloud-bigquery==3.39.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.56.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.0
packaging==25.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
